feat: refactor composables hooks
