+++ src/hooks/useCharacters.ts
@@ -1,10 +1,10 @@
 import { useCharacterStore } from '@/store/characterStore';
 import { useQuery } from '@tanstack/vue-query';
-import { getCharacters } from '../api/characters';
+import { getCharacters, getSingleCharacter } from '../api/characters';
 import { storeToRefs } from 'pinia';
 import { watchEffect } from 'vue';
 
-export default function useCharacters() {
+export const useCharacters = () => {
   const store = useCharacterStore();
   const { characters, currentPage, totalPages } = storeToRefs(store);
 
@@ -25,4 +25,25 @@ export default function useCharacters() {
     totalPages,
     getPage: store.setPageCharacter
   };
-}
+};
+
+export const useCurrentCharacter = (characterId: number) => {
+  const store = useCharacterStore();
+  const { currentCharacter } = storeToRefs(store);
+
+  const { data, isFetching, isError } = useQuery({
+    queryKey: ['character'],
+    queryFn: () => getSingleCharacter(characterId),
+    retry: false
+  });
+
+  watchEffect(() => {
+    if (data.value) store.setCurrentCharacter(data.value);
+  });
+
+  return {
+    currentCharacter,
+    isFetching,
+    isError
+  };
+};
